fix: reorganized code and rename variables
