+++ app/containers/CustomHeader/components/HeaderTitle/styles.ts
@@ -1,0 +1,25 @@
+import { StyleSheet } from 'react-native';
+
+import sharedStyles from '../../../../views/Styles';
+
+export const styles = StyleSheet.create({
+	headerTitleContainer: {
+		flex: 1,
+		alignItems: 'center',
+		justifyContent: 'center'
+	},
+	title: {
+		...sharedStyles.textBold,
+		fontSize: 16,
+		flex: 1,
+		lineHeight: 24,
+		paddingVertical: 6
+	},
+	androidTitle: {
+		...sharedStyles.textBold,
+		fontSize: 16,
+		flex: 1,
+		lineHeight: 24,
+		paddingVertical: 10
+	}
+});
